finish common text input task except optinal icon
