+++ .vscode/settings.json
@@ -21,6 +21,7 @@
               ],
        },
        "cSpell.words": [
+              "autocorrect",
               "roboto"
        ]
 }
